Refactor most factories to be smaller (#230)
### Description of the Change

This PR refactors most of the factory classes to use C-style macros to generate most of the factory methods instead of writing all of them by hand. This approach is much more maintainable and should be backward-compatible.

### Benefits

More maintainable code.

### Possible Drawbacks

This format is harder to document and harder to understand if you just read the header files. I am currently thinking that we can document it using a table for the various parameters.

### Verification Process

This was spot-tested by hand.

### Applicable Issues

Closes #223.