Add a method to APIC to interrupt a movement (#262)
### Description of the Change

This PR adds a method `stop()` to interrupt a motor mid-movement and stop it. Disabling the controller will accomplish this, but then you'll need to manage the disabled lifecycle when making any movements.

### Benefits

A nicer API to stop the controller.

### Possible Drawbacks

Stopping the controller and then disabling/enabling it will cause it to move back to its last set target. This is the intended behavior, but some users might get caught off-guard by it.

### Verification Process

This change was not verified.

### Applicable Issues

Closes #259.