Revamp logger (#289)
### Description of the Change

This PR restructures `Logger` so it's not static and so it is `constexpr` where possible.

### Benefits

The compiler can now optimize out more logger statements when they have been disabled. The logger is now configurable on a per-use basis (and this is reflected in the factories).

### Possible Drawbacks

None.

### Verification Process

This was not verified.

### Applicable Issues

None.