Fix #293 and add tests for it (#319)
### Description of the Change

Currently, when SkidSteerModel's `.arcade` is passed `(-0.0, -1.0)`, it goes backwards. This is not the expected behavior, turning counter-clockwise. This issue is caused by the ability of `std::copysign` to extract the raw sign bit of a `double` number. To fix this, I've changed the comparison of the inputs to the threshold from `<` to `<=`. This will catch -0, and set it to +0.

### Benefits

This will make `.arcade` properly when it's inputs are negated. This is useful for more complex operator control code, like our team's, which has a reverse button.

### Possible Drawbacks

Before, if you passed 0.5 as a threshold, and 0.5 as an input, it would stay 0.5. This change will change that, and make it 0. I believe that this is fine, it is unlikely that this will affect anything in practice.

### Verification Process

I added a test to make sure that a negative 0 as the forward speed does not change the behavior of turning.

### Applicable Issues

Closes #293.