Derive ControllerInput<bool> for AbstractButton (#323)
### Description of the Change

Buttons have a trivial implementation of `ControllerInput<bool>`: `isPressed()`. Adding this implementation allows using buttons (and limit switches, etc.) as inputs to motor controllers in user code.

### Benefits

This allows users of the library to create motor controllers which take boolean sensors as input.

### Possible Drawbacks

N/A

### Verification Process

The code is simple enough that it can be verified visually. This doesn't remove or change functionality, so it cannot introduce any regressions.

### Applicable Issues

N/A