Units: Add getShortUnitName to get string suffix
The function std::string getShortUnitName(QType q) in
include/okapi/api/units/RQuantityName.hpp returns the a string of the
matching unit.

For example, a call such as `getShortUnitName(1_m)` will return "m",
1_rad will return "rad". A value that doesn't exactly (within 0.01%) map
to a unit such as 1.01_m will throw an error.

This is currently only implemented for QLength and QAngle because that's
what's needed for Odometry. It should be trivial to add other types to
the map if the need arises in the future.